+++ src/components/LineBetween/utils.js
@@ -1,11 +1,18 @@
-/** @param {string | Element} selector */
-export const getElement = selector => {
+/**
+ * @param {string | Element} selector
+ * @param {string | Element} [container]
+ *
+ * @return {HTMLElement | null}
+ */
+export const getElement = (selector, container) => {
   if (selector instanceof Element) {
     return selector;
   }
 
   try {
-    return document.querySelector(selector);
+    const containerElem = container ? getElement(container) : document.body;
+
+    return containerElem.querySelector(selector);
   } catch (error) {
     return null;
   }