+++ src/components/JobLogs/jobLogPropTypes.js
@@ -2,8 +2,9 @@ import PropTypes from 'prop-types';
 
 export const jobLogPropTypes = PropTypes.shape({
   fileName: PropTypes.string.isRequired,
-  jobProfileName: PropTypes.string.isRequired,
   jobExecutionHrId: PropTypes.string.isRequired,
+  jobProfileName: PropTypes.string.isRequired,
+  totalRecords: PropTypes.number.isRequired,
   completedDate: PropTypes.string.isRequired,
   runBy: PropTypes.shape({
     firstName: PropTypes.string.isRequired,