+++ test/bigtest/tests/logs-list-test.js
@@ -50,57 +50,33 @@ describe('Logs list', () => {
 
     describe('compares fields as numbers', () => {
       beforeEach(async () => {
-        await logsList.headers(2).click();
+        await logsList.headers(1).click();
       });
 
       it('ascending', () => {
-        expect(getCellContent(0, 2)).to.equal('01');
-        expect(getCellContent(1, 2)).to.equal('02');
-        expect(getCellContent(2, 2)).to.equal('03');
+        expect(getCellContent(0, 1)).to.equal('01');
+        expect(getCellContent(1, 1)).to.equal('02');
+        expect(getCellContent(2, 1)).to.equal('03');
       });
 
       describe('and', () => {
         beforeEach(async () => {
-          await logsList.headers(2).click();
+          await logsList.headers(1).click();
         });
 
         it('descending', () => {
-          expect(getCellContent(0, 2)).to.equal('03');
-          expect(getCellContent(1, 2)).to.equal('02');
-          expect(getCellContent(2, 2)).to.equal('01');
+          expect(getCellContent(0, 1)).to.equal('03');
+          expect(getCellContent(1, 1)).to.equal('02');
+          expect(getCellContent(2, 1)).to.equal('01');
         });
       });
     });
 
     describe('compares fields as dates', () => {
       it('descending', () => {
-        expect(getCellContent(0, 3)).to.equal('11/11/2018, 2:10 PM');
-        expect(getCellContent(1, 3)).to.equal('11/5/2018, 2:21 PM');
-        expect(getCellContent(2, 3)).to.equal('11/5/2018, 1:08 PM');
-      });
-
-      describe('and', () => {
-        beforeEach(async () => {
-          await logsList.headers(3).click();
-        });
-
-        it('ascending', () => {
-          expect(getCellContent(0, 3)).to.equal('11/5/2018, 1:08 PM');
-          expect(getCellContent(1, 3)).to.equal('11/5/2018, 2:21 PM');
-          expect(getCellContent(2, 3)).to.equal('11/11/2018, 2:10 PM');
-        });
-      });
-    });
-
-    describe('compares by `runBy` field', () => {
-      beforeEach(async () => {
-        await logsList.headers(4).click();
-      });
-
-      it('ascending', () => {
-        expect(getCellContent(0, 4)).to.equal('Elliot Lane');
-        expect(getCellContent(1, 4)).to.equal('Jay Morrowitz');
-        expect(getCellContent(2, 4)).to.equal('Ozzy Campenshtorm');
+        expect(getCellContent(0, 4)).to.equal('11/11/2018, 2:10 PM');
+        expect(getCellContent(1, 4)).to.equal('11/5/2018, 2:21 PM');
+        expect(getCellContent(2, 4)).to.equal('11/5/2018, 1:08 PM');
       });
 
       describe('and', () => {
@@ -108,10 +84,34 @@ describe('Logs list', () => {
           await logsList.headers(4).click();
         });
 
+        it('ascending', () => {
+          expect(getCellContent(0, 4)).to.equal('11/5/2018, 1:08 PM');
+          expect(getCellContent(1, 4)).to.equal('11/5/2018, 2:21 PM');
+          expect(getCellContent(2, 4)).to.equal('11/11/2018, 2:10 PM');
+        });
+      });
+    });
+
+    describe('compares by `runBy` field', () => {
+      beforeEach(async () => {
+        await logsList.headers(5).click();
+      });
+
+      it('ascending', () => {
+        expect(getCellContent(0, 5)).to.equal('Elliot Lane');
+        expect(getCellContent(1, 5)).to.equal('Jay Morrowitz');
+        expect(getCellContent(2, 5)).to.equal('Ozzy Campenshtorm');
+      });
+
+      describe('and', () => {
+        beforeEach(async () => {
+          await logsList.headers(5).click();
+        });
+
         it('descending', () => {
-          expect(getCellContent(0, 4)).to.equal('Ozzy Campenshtorm');
-          expect(getCellContent(1, 4)).to.equal('Jay Morrowitz');
-          expect(getCellContent(2, 4)).to.equal('Elliot Lane');
+          expect(getCellContent(0, 5)).to.equal('Ozzy Campenshtorm');
+          expect(getCellContent(1, 5)).to.equal('Jay Morrowitz');
+          expect(getCellContent(2, 5)).to.equal('Elliot Lane');
         });
       });
     });