+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * Integrate updated MultiColumnList component (UIDATIMP-276)
 * Create Section component (UIDATIMP-278)
 * Create and integrate RecordTypesSelect component (UIDATIMP-244)
+* Add a Records column to a job logs table, rename Import ID column to ID, swap columns and change width of panes (UIDATIMP-260)
+* Fix save match profile (UIDATIMP-282)
 
 ## [1.4.0](https://github.com/folio-org/ui-data-import/tree/v1.4.0) (2019-08-02)
 * Create data import settings page's 3rd pane for action profiles (UIDATIMP-8)